agents/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.45.0
langgraph==1.1.9
langchain==1.2.15
langchain-openai==1.1.16
langchain-anthropic==1.4.1
python-dotenv==1.2.2
pydantic==2.13.3
httpx-sse>=0.4.0
